Resolve "EP-57: Create unit tests for User, Person, and Friend"

Closes #57

See merge request rybolzde/event-planner!8

+++ oneplan/oneplan/src/main/java/fit/biesp/oneplan/service/FriendService.java
@@ -1,6 +1,7 @@
 package fit.biesp.oneplan.service;
 
 import fit.biesp.oneplan.entity.FriendEntity;
+import fit.biesp.oneplan.exception.PersonNotFoundException;
 import fit.biesp.oneplan.model.FriendModel;
 import fit.biesp.oneplan.repository.FriendRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,33 +14,38 @@ public class FriendService {
     @Autowired
     private FriendRepository friendRepository;
 
-    public FriendEntity create(FriendEntity friend){
+    public FriendEntity create(FriendEntity friend) {
         return friendRepository.save(friend);
     }
 
-    public FriendModel addFriend(FriendModel friendModel){
+    public FriendModel addFriend(FriendModel friendModel) {
         return FriendModel.toModel(friendRepository.save(FriendModel.fromModel(friendModel)));
     }
 
-    public FriendModel getFriend(String nickname){
+    public FriendModel getFriend(String nickname) throws PersonNotFoundException {
         FriendEntity friend = friendRepository.findByNickname(nickname);
+        if (friend == null) throw new PersonNotFoundException("");
         return FriendModel.toModel(friend);
     }
 
-    public FriendModel updateFriend(FriendModel friendModel, String nickname){
+    public FriendModel updateFriend(FriendModel friendModel, String nickname) throws PersonNotFoundException {
         FriendEntity friend = friendRepository.findByNickname(nickname);
+        if (friend == null) throw new PersonNotFoundException("");
         friend.setNickname(friendModel.getNickname());
         friend.setEmail(friendModel.getEmail());
 
         return FriendModel.toModel(friendRepository.save(friend));
     }
 
-    public Long delete(String nickname){
+    public Long delete(String nickname) throws PersonNotFoundException {
         FriendEntity friend = friendRepository.findByNickname(nickname);
+        if (friend == null) throw new PersonNotFoundException("");
         Long id = friend.getId();
         friendRepository.deleteById(id);
         return id;
     }
 
-    public List<FriendEntity> findAllByUserId(Integer userId){ return friendRepository.findAllByUserId(userId);}
+    public List<FriendEntity> findAllByUserId(Integer userId) {
+        return friendRepository.findAllByUserId(userId);
+    }
 }
